fix(dashboard): update errors to deploy to Vercel
